+++ unlock-app/src/components/interface/locks/Settings/index.tsx
@@ -7,7 +7,7 @@ import { useLockManager } from '~/hooks/useLockManager'
 import { useAuth } from '~/contexts/AuthenticationContext'
 import { addressMinify } from '~/utils/strings'
 import { SettingHeader } from './elements/SettingHeader'
-import { useQuery } from '@tanstack/react-query'
+import { useQueries, useQuery } from '@tanstack/react-query'
 import { useWeb3Service } from '~/utils/withWeb3Service'
 import { SettingGeneral } from './elements/SettingGeneral'
 import { SettingMisc } from './elements/SettingMisc'
@@ -76,6 +76,22 @@ const LockSettingsPage = ({ lockAddress, network }: LockSettingsPageProps) => {
       refetchInterval: 1000,
     }
   )
+
+  const [{ data: publicLockLatestVersion }, { data: publicLockVersion }] =
+    useQueries({
+      queries: [
+        {
+          queryKey: ['publicLockLatestVersion', network],
+          queryFn: async () =>
+            await web3Service.publicLockLatestVersion(network),
+        },
+        {
+          queryKey: ['publicLockVersion', lockAddress, network],
+          queryFn: async () =>
+            await web3Service.publicLockVersion(lockAddress, network),
+        },
+      ],
+    })
 
   const isLoading = isLoadingLock || isLoadingManager
 
@@ -159,7 +175,8 @@ const LockSettingsPage = ({ lockAddress, network }: LockSettingsPageProps) => {
           network={network}
           isManager={isManager}
           isLoading={isLoading}
-          lock={lock}
+          publicLockLatestVersion={publicLockLatestVersion}
+          publicLockVersion={publicLockVersion}
         />
       ),
       sidebar: (